[Refactor] Restructure the apps; folder structure to organise files
