+++ public/auth/signin/index.html
@@ -11,7 +11,7 @@
 
 <body>
   <div class="container container--auth">
-    <div class="card card--auth card--login">
+    <div class="card card--login">
       <div class="card__header">
         <span class="brand">
           <a href="../../index.html" class="link">M-tracker</a>
