+++ public/dashboard/details/index.html
@@ -81,11 +81,19 @@
           </ul>
         </nav>
       </header>
-      <main class="dashboard__content-section" id="content"></main>
+      <main class="dashboard__content-section">
+        <div class="fab white mb-md" id="back">
+          <svg class="icon">
+            <use xlink:href="../../assets/icons/sprite.svg#icon-arrow-left"></use>
+          </svg>
+        </div>
+
+        <div id="content"></div>
+      </main>
     </div>
   </div>
 
-  <script src="../../dist/details.bundle.js"></script>
+  <script src="../../dist/admin_details.bundle.js"></script>
 </body>
 
 </html>
