+++ src/components/Admin/Dashboard.js
@@ -1,11 +1,11 @@
-import { Component } from "./App";
-import { dashboardNodes } from "./nodes";
-import api from "./api";
-import { redirect } from "./routes";
+import { Component } from "../utils/App";
+import { dashboardNodes } from "../utils/nodes";
+import api from "../utils/api";
+import { redirect } from "../utils/routes";
 
 class Dashboard extends Component {
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
     this.state = {
       isAuthenticated: false,
       isFetching: false,
@@ -22,6 +22,7 @@ class Dashboard extends Component {
     };
     this.elems = dashboardNodes();
     this.handleTokenUpdate();
+    this.fetchRequests();
   }
 
   handleTokenUpdate() {
@@ -98,11 +99,9 @@ class Dashboard extends Component {
                 this.elems.content.innerHTML =
                   tableHeader +
                   data.requests.map(
-                    request =>
-                      `<a href="/dashboard/details/?${
-                        request.public_id
-                      }" class="table--link">
-                    <div class="hr row--table">
+                    request => `
+                    <a href="details/?${request.public_id}" class="table--link">
+                      <div class="hr row--table">
                       <div class="table__data table__data--md">
                         ${
                           request.title.length > 22
@@ -138,6 +137,4 @@ class Dashboard extends Component {
     });
   }
 }
-
 const dashboard = new Dashboard();
-dashboard.fetchRequests();
