+++ src/components/Admin/RequestDetails.js
@@ -1,0 +1,3 @@
+import Details from "../Base/Details";
+
+new Details({ uri: "/request/" });
