+++ src/components/Guest/Signup.js
@@ -1,30 +1,33 @@
-import { signinNodes } from "./nodes";
-import { redirect } from "./routes";
-import { Component } from "./App";
-import api from "./api";
+import api from "../utils/api";
+import { Component } from "../utils/App";
+import { redirect } from "../utils/routes";
+import { signupNodes } from "../utils/nodes";
 
-class Signin extends Component {
-  constructor() {
-    super();
+class Signup extends Component {
+  constructor(props) {
+    super(props);
     this.state = {
       data: {},
       errors: {},
-      success: null
+      success: null,
+      isFetching: false
     };
-    this.elements = signinNodes();
-    this.submit = document.forms.signin.elements.submit;
+    this.elements = signupNodes();
+    this.submit = document.forms.signup.elements.submit;
     this.handleChange = this.handleChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
+    this.handleChange();
+    this.handleSubmit();
     this.showLoaders();
   }
 
   showLoaders() {
     setInterval(() => {
       if (this.state.isFetching) {
-        this.submit.innerHTML = "Signing in...";
+        this.submit.innerHTML = "Signing up...";
         this.submit.setAttribute("disabled", "disabled");
       } else {
-        this.submit.innerHTML = "Sign in";
+        this.submit.innerHTML = "Sign up";
         this.submit.removeAttribute("disabled");
       }
     }, 0);
@@ -46,9 +49,9 @@ class Signin extends Component {
     this.submit.addEventListener("click", e => {
       e.preventDefault();
       this.setState({ isFetching: true });
-      console.log(this.state.data);
+      console.log(this.state);
       api
-        .post("/users/auth/signin/", this.state.data)
+        .post("/users/auth/signup/", this.state.data)
         .then(res => {
           this.setState({ success: res.ok });
           return res.json();
@@ -56,24 +59,23 @@ class Signin extends Component {
         .then(data => {
           this.setState({ isFetching: false });
           if (this.state.success) {
-            localStorage.setItem("token", data.access_token);
-            redirect("/requests/");
+            localStorage.setItem("success", data.message);
+            redirect("/auth/signin/");
           } else {
-            Object.entries(data).map(val => {
-              errors[val[0]] = val[1];
-              this.setState({ errors });
+            Object.values(data).map(value => {
+              Object.entries(value).map(val => {
+                errors[val[0]] = val[1];
+                this.setState({ errors });
+              });
             });
-            console.log(this.state.errors);
           }
         })
         .catch(err => {
-          this.setState({ isFetching: true });
+          this.setState({ isFetching: false });
           console.log(err.message);
         });
     });
   }
 }
 
-const signin = new Signin();
-signin.handleChange();
-signin.handleSubmit();
+const signup = new Signup();
