--- src/components/routes.js
@@ -1,3 +1,0 @@
-const baseRoute = "http://127.0.0.1:8080";
-
-export const redirect = path => window.location.replace(`${baseRoute}${path}`);
